move errors to their appropriate places
